code file
This project implements a Support Vector Machine (SVM) classifier to predict whether a patient has diabetes based on various health metrics. The model was trained on the Pima Indians Diabetes Dataset, which contains 768 samples with 8 medical features and a binary outcome (1 for diabetic, 0 for non-diabetic).The model is trained on various types of patients body parameters like Insulin level, Blood glucose level, BMI level etc. Once the trained on input data then we can feed unseen data to the our system and our system predicts whether a patients has diabetes or not based on input data feed to the system.

This system is achieved 78.8% accuracy on training data and 76.6% accuracy on test data.